WIP of revamping users profile page.
This commit takes away the option of adding a membership from your profile page.
This is in anticipation of the searchable directory feature that is coming back.

This commit also adds a little style to the list of memberships, although we still need
to design a way to figure out how to show membership status