comment out views table migration.
Our production db is a clone of the old db which already has this table.
It was added in a previous version. I manually deleted the entry of this
migration from schema_migrations because there are migrations before
this that have not been run. But since thist able already exits,
this migration will not run. So I am commenting this out for now
so that migrations will run and update the schema_migrations table
and our schema.rb file. From now on, we should use schema:load instead
of cloning the db, and then find some sane way to port over production
data.